+++ lib/resolution/resolver.js
@@ -8,12 +8,14 @@ const { StructuredInlineDeclarationResolver } = require('../components/inline')
 const { isInlineEnumType, propertyToInlineEnumName } = require('../components/enum')
 const { isReferenceType } = require('../components/reference')
 const { isEntity, getMaxCardinality } = require('../csn')
-const { baseDefinitions } = require('../components/basedefs')
+const { getBaseDefinitions } = require('../components/basedefs')
 const { BuiltinResolver } = require('./builtin')
 const { LOG } = require('../logging')
 const { last } = require('../components/identifier')
 const { getPropertyModifiers } = require('../components/property')
 const { configuration } = require('../config')
+
+const baseDefinitions = getBaseDefinitions()
 
 /** @typedef {import('../visitor').Visitor} Visitor */
 /** @typedef {import('../typedefs').resolver.CSN} CSN */
