+++ lib/visitor.js
@@ -10,13 +10,15 @@ const { docify, createPromiseOf, createUnionOf, createKeysOf, createElementsOf, 
 const { csnToEnumPairs, propertyToInlineEnumName, isInlineEnumType, stringifyEnumType } = require('./components/enum')
 const { isReferenceType } = require('./components/reference')
 const { empty } = require('./components/typescript')
-const { baseDefinitions } = require('./components/basedefs')
+const { getBaseDefinitions } = require('./components/basedefs')
 const { EntityRepository, asIdentifier } = require('./resolution/entity')
 const { last } = require('./components/identifier')
 const { getPropertyModifiers } = require('./components/property')
 const { configuration } = require('./config')
 const { createMember } = require('./components/class')
 const { overrideNameProperty } = require('./printers/javascript')
+
+const baseDefinitions = getBaseDefinitions()
 
 /** @typedef {import('./file').File} File */
 /** @typedef {import('./typedefs').visitor.Context} Context */
